requirements.txt
docx-mailmerge==0.5.0
python-decouple==3.7
shodan==1.28.0